refacto: extract same logic in all converters
