+++ converter/converter/versions/create_case_health/create_case_health_converter.py
@@ -3,35 +3,16 @@ import copy
 
 from converter.utils import add_to_medical_notes, delete_paths, get_field_value, is_field_completed, map_to_new_value
 from converter.versions.base_message_converter import BaseMessageConverter
+from converter.versions.conversion_mixin import ConversionMixin
 from converter.versions.create_case_health.v1_v2.constants import V1V2Constants
 from converter.versions.create_case_health.v2_v3.constants import V2V3Constants
 from converter.versions.utils import reverse_map_to_new_value, switch_field_name, validate_diagnosis_code
 
 
-class CreateHealthCaseConverter(BaseMessageConverter):
+class CreateHealthCaseConverter(BaseMessageConverter, ConversionMixin):
     @staticmethod
     def get_message_type() -> str:
         return "createCaseHealth"
-
-    @classmethod
-    def copy_input_content(cls, input_json: Dict[str, Any]) -> Dict[str, Any]:
-        # Create independent envelope copy without use case for output
-        output_json = copy.deepcopy(input_json)
-        if cls.get_message_type() not in input_json.get('content', [{}])[0].get('jsonContent', {}).get('embeddedJsonContent', {}).get('message', {}):
-            raise ValueError(f"Input JSON must contain {cls.get_message_type()} key")
-        del output_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()]
-        return output_json
-
-    @classmethod
-    def copy_input_use_case_content(cls, input_json: Dict[str, Any]) -> Dict[str, Any]:
-        # Create independent use case copy for output
-        input_use_case_json = input_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()]
-        return copy.deepcopy(input_use_case_json)
-
-    @classmethod
-    def format_output_json(cls, output_json: Dict[str, Any], output_use_case_json: Dict[str, Any]) -> Dict[str, Any]:
-        output_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()] = output_use_case_json
-        return output_json
 
     @classmethod
     def convert_v1_to_v2(cls, input_json: Dict[str, Any]) -> Dict[str, Any]:
@@ -129,10 +110,6 @@ class CreateHealthCaseConverter(BaseMessageConverter):
     def convert_v2_to_v3(cls, input_json: Dict[str, Any]) -> Dict[str, Any]:
         output_json = cls.copy_input_content(input_json)
         output_use_case_json = cls.copy_input_use_case_content(input_json)
-
-        # Create independent use case copy for output
-        input_use_case_json = input_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()]
-        output_use_case_json = copy.deepcopy(input_use_case_json)
 
         map_to_new_value(output_use_case_json,'$.interventionType', V2V3Constants.V2_TO_V3_INTERVENTION_TYPE_MAPPING)
         map_to_new_value(output_use_case_json,'$.qualification.origin', V2V3Constants.V2_TO_V3_QUALIFICATION_ORIGIN_MAPPING)
