+++ converter/converter/versions/resources_request/resources_request_converter.py
@@ -3,33 +3,14 @@ from typing import Dict, Any
 
 from converter.utils import map_to_new_value
 from converter.versions.base_message_converter import BaseMessageConverter
+from converter.versions.conversion_mixin import ConversionMixin
 from converter.versions.resources_request.resources_request_constants import ResourcesRequestConstants
 from converter.versions.utils import reverse_map_to_new_value
 
-class ResourcesRequestConverter(BaseMessageConverter):
+class ResourcesRequestConverter(BaseMessageConverter, ConversionMixin):
     @staticmethod
     def get_message_type():
         return "resourcesRequest"
-
-    @classmethod
-    def copy_input_content(cls, input_json: Dict[str, Any]) -> Dict[str, Any]:
-        # Create independent envelope copy without use case for output
-        output_json = copy.deepcopy(input_json)
-        if cls.get_message_type() not in input_json.get('content', [{}])[0].get('jsonContent', {}).get('embeddedJsonContent', {}).get('message', {}):
-            raise ValueError(f"Input JSON must contain {cls.get_message_type()} key")
-        del output_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()]
-        return output_json
-
-    @classmethod
-    def copy_input_use_case_content(cls, input_json: Dict[str, Any]) -> Dict[str, Any]:
-        # Create independent use case copy for output
-        input_use_case_json = input_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()]
-        return copy.deepcopy(input_use_case_json)
-
-    @classmethod
-    def format_output_json(cls, output_json: Dict[str, Any], output_use_case_json: Dict[str, Any]) -> Dict[str, Any]:
-        output_json['content'][0]['jsonContent']['embeddedJsonContent']['message'][cls.get_message_type()] = output_use_case_json
-        return output_json
 
     @classmethod
     def convert_v1_to_v2(cls, input_json) -> Dict[str, Any]:
